refactor: reuse contacts for dynamic
